controllers-pages addiction

+++ service-layer/src/main/java/com/gmail/sasha/myproject/service/service/ItemDiscountService.java
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 import java.util.List;
 
 public interface ItemDiscountService {
-    void assignToRangeOfItemsCorrespondingDiscounts(int minItemPrice, int maxItemPrice, BigDecimal discountValue);
+    void assignToRangeOfItemsCorrespondingDiscounts(BigDecimal minItemPrice, BigDecimal maxItemPrice, BigDecimal discountValue);
 
     List<ItemWithDiscountedPrice> showItemsWithDiscountedPrice();
 }
