+++ service-layer/src/main/java/com/gmail/sasha/myproject/service/service/ItemService.java
@@ -2,6 +2,7 @@ package com.gmail.sasha.myproject.service.service;
 
 import com.gmail.sasha.myproject.service.model.ItemDTO;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ItemService {
@@ -10,7 +11,7 @@ public interface ItemService {
 
     List<ItemDTO> getAllItems();
 
-    List<ItemDTO> getItemsInPriceRange(int from, int to);
+    List<ItemDTO> getItemsInPriceRange(BigDecimal from, BigDecimal to);
 
     ItemDTO findById(ItemDTO itemDTO);
